altos/telegps: Create new flight if current flight is erased
telegps is unique in that USB may be connected while a flight is
active and sensible things should happen. If a flight is being
recorded and gets erased, then a new flight should be started.

This is done by hooking in the flight erase code and calling out to
the tracker code to figure out whether to switch to a new flight or not.

Signed-off-by: Keith Packard <keithp@keithp.com>